refactor: update components file trees
