+++ src/components/Order/index.ts
@@ -1,0 +1,3 @@
+import Orders from './Orders';
+
+export default Orders;
