+++ src/components/PasswordReset/index.ts
@@ -1,0 +1,4 @@
+export { default as NewPassword } from './NewPassword';
+export { default as ResetPassword } from './ResetPassword';
+export { default as ResetSuccessfully } from './ResetSuccessfully';
+export { default as VerifyEmail } from './VerifyEmail';
