+++ src/components/Product/index.ts
@@ -1,0 +1,3 @@
+export { default as Products } from './Products';
+export { default as CreateProduct } from './CreateProduct';
+export { default as ProductDetails } from './ProductDetails';
