+++ src/pages/products/[id].tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ProductDetails from '@components/ProductDetails';
+import { ProductDetails } from '@components/Product';
 
 export default function SingleProductPage() {
   return <ProductDetails />;
