+++ src/pages/products/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Products from '@components/Products';
+import { Products } from '@components/Product';
 
 export default function ProductPage() {
   return <Products />;
