+++ src/pages/reset-password.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
 
-import NewPassword from '@components/PasswordReset/NewPassword';
-import ResetPassword from '@components/PasswordReset/ResetPassword';
-import ResetSuccessfully from '@components/PasswordReset/ResetSuccessfully';
-import VerifyEmail from '@components/PasswordReset/VerifyEmail';
+import {
+  NewPassword,
+  ResetPassword,
+  ResetSuccessfully,
+  VerifyEmail,
+} from '@components/PasswordReset';
 
 const ResetPasswordPage = () => <ResetSuccessfully />;
