+++ src/pages/sell.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import CreateProduct from '@components/CreateProduct';
+import { CreateProduct } from '@components/Product';
 
 export default function SellPage() {
   return <CreateProduct />;
